refactor: rename directory to identity-result
