+++ apps/explorer/src/pages/trust-framework/identity-result/DidResult.tsx
