+++ apps/explorer/src/pages/trust-framework/identity-result/didServiceHelper.ts
