+++ apps/explorer/src/pages/trust-framework/identity-result/headerMetadataHelper.ts
