+++ apps/explorer/src/pages/trust-framework/identity-result/helper.ts
