+++ apps/explorer/src/pages/trust-framework/identity-result/hooks/useGetControllerObjects.ts
