+++ apps/explorer/src/pages/trust-framework/identity-result/types/index.ts
