+++ apps/explorer/src/pages/trust-framework/identity-result/views/SideBySidePanelsView.tsx
